Modified repairtask to allow for dynamic amount of repairamount. Rope-art, TaskBoard gameobject (table), smaller path grid
Added some roped-off art, not yet dynamically used. Supposed to show up for planned maintenance.

Worker now has to go to Maintenance Manager to request a task and to finalize a task.

Changed machine generate repair task behavior:

Earlier, the task expected worker to fully repair machine before setting it to stopped (ready for operator to set running).

Now it only needs to be repaired a specific a mount (% of max condition) before worker sets it to repair. Of course this means that the machine is not going to run for as long before breaking down again, but perhaps it can get production up and running and then at scheduled maintenance, a full repair will be performed.

Modified the Generate Repair Task and the Repair task to allow for a required int repairamount which is kept track of by the task. The machine evaluates how much to be repaired.